+++ NEWS.md
@@ -1,3 +1,16 @@
+# gMOIP 1.4.3
+
+* Added support for rays in different directions.
+
+# gMOIP 1.4.2
+
+* Classification of nondominated points.
+
+# gMOIP 1.4.0
+
+* Added support for plotting rectangles, points, planes, hulls in 3D. 
+* Nondominated points in 3D can now be generated.
+
 # gMOIP 1.3.0
 
 * Added support for 3D plots using RGL.